+++ demisto_sdk/commands/init/tests/initiator_test.py
@@ -24,6 +24,8 @@ PACK_TAGS = 'Tag1,Tag2'
 PACK_GITHUB_USERS = ''
 INTEGRATION_NAME = 'IntegrationName'
 SCRIPT_NAME = 'ScriptName'
+DEFAULT_INTEGRATION = 'BaseIntegration'
+DEFAULT_SCRIPT = 'BaseScript'
 
 
 @pytest.fixture
@@ -332,7 +334,7 @@ def test_get_remote_templates__valid(mocker, initiator):
     mocker.patch.object(tools, 'get_remote_file', return_value=b'Test im in file')
     initiator.full_output_path = PACK_NAME
     os.makedirs(PACK_NAME, exist_ok=True)
-    res = initiator.get_remote_templates(['Test.py'])
+    res = initiator.get_remote_templates(['Test.py'], dir=DIR_NAME)
     file_path = os.path.join(PACK_NAME, 'Test.py')
     with open(file_path, 'r') as f:
         file_content = f.read()
@@ -359,7 +361,7 @@ def test_get_remote_templates__invalid(mocker, initiator):
     mocker.patch.object(tools, 'get_remote_file', return_value={})
     initiator.full_output_path = PACK_NAME
     os.makedirs(PACK_NAME, exist_ok=True)
-    res = initiator.get_remote_templates(['Test.py'])
+    res = initiator.get_remote_templates(['Test.py'], dir=DIR_NAME)
 
     assert not res
 
@@ -380,7 +382,7 @@ def test_integration_init(initiator, tmpdir):
     Then
         - Ensure the function's return value is True
         - Ensure integration directory with the desired integration name is created successfully.
-        - Ensure integration directory contain all files.
+        - Ensure integration directory contain all files of the Boilerplate template.
     """
     temp_pack_dir = os.path.join(tmpdir, PACK_NAME)
     os.makedirs(temp_pack_dir, exist_ok=True)
@@ -388,14 +390,53 @@ def test_integration_init(initiator, tmpdir):
     initiator.output = temp_pack_dir
     initiator.dir_name = INTEGRATION_NAME
     initiator.is_integration = True
+    initiator.template = DEFAULT_INTEGRATION
 
     integration_path = os.path.join(temp_pack_dir, INTEGRATION_NAME)
     res = initiator.integration_init()
     integration_dir_files = {file for file in listdir(integration_path)}
     expected_files = {
-        "Pipfile", "Pipfile.lock", f"{INTEGRATION_NAME}.py",
+        "Pipfile", "Pipfile.lock", "command_examples", "test_data", "README.md", f"{INTEGRATION_NAME}.py",
         f"{INTEGRATION_NAME}.yml", f"{INTEGRATION_NAME}_description.md", f"{INTEGRATION_NAME}_test.py",
-        f"{INTEGRATION_NAME}_image.png", "test_data"
+        f"{INTEGRATION_NAME}_image.png"
+    }
+
+    assert res
+    assert os.path.isdir(integration_path)
+    assert expected_files == integration_dir_files
+
+
+def test_template_integration_init(initiator, tmpdir):
+    """
+    Tests `integration_init` function with a given integration template name.
+
+    Given
+        - Inputs to init integration in a given output.
+        - An integration template - HelloWorld.
+
+    When
+        - Running the init command.
+
+    Then
+        - Ensure the function's return value is True
+        - Ensure integration directory with the desired integration name is created successfully.
+        - Ensure integration directory contains all the files of the template integration.
+    """
+    temp_pack_dir = os.path.join(tmpdir, PACK_NAME)
+    os.makedirs(temp_pack_dir, exist_ok=True)
+
+    initiator.output = temp_pack_dir
+    initiator.dir_name = INTEGRATION_NAME
+    initiator.is_integration = True
+    initiator.template = 'HelloWorld'
+
+    integration_path = os.path.join(temp_pack_dir, INTEGRATION_NAME)
+    res = initiator.integration_init()
+    integration_dir_files = {file for file in listdir(integration_path)}
+    expected_files = {
+        "Pipfile", "Pipfile.lock", "README.md", f"{INTEGRATION_NAME}.py",
+        f"{INTEGRATION_NAME}.yml", f"{INTEGRATION_NAME}_description.md", f"{INTEGRATION_NAME}_test.py",
+        f"{INTEGRATION_NAME}_image.png", "test_data", "command_examples"
     }
 
     assert res
@@ -421,6 +462,7 @@ def test_script_init(initiator, tmpdir):
     temp_pack_dir = os.path.join(tmpdir, PACK_NAME)
     os.makedirs(temp_pack_dir, exist_ok=True)
 
+    initiator.template = DEFAULT_SCRIPT
     initiator.dir_name = SCRIPT_NAME
     initiator.output = temp_pack_dir
     script_path = os.path.join(temp_pack_dir, SCRIPT_NAME)
@@ -430,4 +472,5 @@ def test_script_init(initiator, tmpdir):
 
     assert res
     assert os.path.isdir(script_path)
-    assert {f"{SCRIPT_NAME}.py", f"{SCRIPT_NAME}.yml", f"{SCRIPT_NAME}_test.py"} == script_dir_files
+    assert {f"{SCRIPT_NAME}.py", f"{SCRIPT_NAME}.yml", f"{SCRIPT_NAME}_test.py",
+            "README.md", "test_data"} == script_dir_files