+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 * The lint command will now ignore the unsubscriptable-object (E1136) pylint error in dockers based on python 3.9 - this will be removed once a new pylint version is released.
 * Added an option for **format** to run on a whole pack.
 * Added new validation of unimplemented commands from yml in the code to `XSOAR-linter`.
+* Fixed an issue where Auto-Extract fields were only checked for newly added incident types in the **validate** command.
 
 # 1.2.13
 * Added new validation of indicators usage in CommandResults to `XSOAR-linter`.