+++ CHANGELOG.md
@@ -1,4 +1,12 @@
 # Changelog
+* SDK repository is now mypy check_untyped_defs complaint.
+* The lint command will now ignore the unsubscriptable-object (E1136) pylint error in dockers based on python 3.9 - this will be removed once a new pylint version is released.
+* Added an option for **format** to run on a whole pack.
+* Added new validation of unimplemented commands from yml in the code to `XSOAR-linter`.
+* Fixed an issue where Auto-Extract fields were only checked for newly added incident types in the **validate** command.
+* Updated **demisto-sdk init** command to no longer create `created` field in pack_metadata file
+
+# 1.2.13
 * Added new validation of indicators usage in CommandResults to `XSOAR-linter`.
 * Running **demisto-sdk lint** will automatically run on changed files (same behavior as the -g flag).
 * Removed supported version message from the documentation when running **generate_docs**.
@@ -11,8 +19,6 @@
 * Added validations that checks in read-me for empty sections or leftovers from the auto generated read-me that should be changed.
 * Added new code validation for *NotImplementedError* to raise a warning in `XSOAR-linter`.
 * Added validation for support types in the pack metadata file.
-* Updated **demisto-sdk init** command to no longer create `created` field in pack_metadata file
-
 
 # 1.2.12
 * Bandit now reports also on medium severity issues.